odometry overlow & position publishing

+++ demoRMR/odometry.cpp
@@ -11,19 +11,8 @@ Odometry::Odometry()
 
 void Odometry::update(TKobukiData robotData)
 {
-    // TODO: static cast:
-
-    // uint16_t cur  = static_cast<uint16_t>(10);
-    // uint16_t prev = static_cast<uint16_t>(65530);
-
-    // int16_t delta = static_cast<int16_t>(cur - prev);
-
-    // std::cout << "cur: " << cur << ", prev: " << prev << "\n";
-    // std::cout << "delta: " << delta << "\n";
-    // return 0;
-
-    int deltaLeft = (signed short) robotData.EncoderLeft - (signed short) _encoderLeftPrev;
-    int deltaRight = (signed short) robotData.EncoderRight - (signed short) _encoderRightPrev;
+    int16_t deltaLeft = static_cast<int16_t>(static_cast<uint16_t>(robotData.EncoderLeft) - static_cast<uint16_t>(_encoderLeftPrev));
+    int16_t deltaRight = static_cast<int16_t>(static_cast<uint16_t>(robotData.EncoderRight) - static_cast<uint16_t>(_encoderRightPrev));
     _encoderLeftPrev = robotData.EncoderLeft;
     _encoderRightPrev = robotData.EncoderRight;
 
@@ -45,7 +34,6 @@ void Odometry::update(TKobukiData robotData)
     _posX += deltaDist * cos(_rot);
     _posY += deltaDist * sin(_rot);
 
-    //std::cout << "Odometry update called encoder data: " << robotData.EncoderLeft << std::endl;
     std::cout << "Odometry: x = " << _posX << ", y = " << _posY << ", rot = " << _rot << std::endl;
 }
